decomp: HolmesClient — HolmesClientInitOpcode 94.3% → 99.7%, cast fix for cmpw
InitOpcode improvement (94.3% → 99.7%) from gRequests moved after gStreamBuffer
in declaration order (already committed). HolmesFlushStreamBuffer: cast size() to
int so comparison emits cmpw (signed) instead of cmplw (unsigned).

HolmesClientClose and HolmesClientTruncate are at_limit (~94%) due to BSS offset
mismatch from EndCmd's GlobalLink/storage being non-consecutive in the target.

+++ src/system/os/HolmesClient.cpp
@@ -114,7 +114,7 @@ namespace {
     }
 
     void HolmesFlushStreamBuffer() {
-        if (gStreamBuffer->mBuffer.size() > 0x2000d)
+        if ((int)gStreamBuffer->mBuffer.size() > 0x2000d)
             gRealMaxBufferSize = gStreamBuffer->mBuffer.size();
         gHolmesStream->Write(&gStreamBuffer->mBuffer[0], gStreamBuffer->mBuffer.size());
         gStreamBuffer->Seek(0, BinStream::kSeekEnd);
